+++ cli/main.go
@@ -1,0 +1,7 @@
+package main
+
+import "github.com/c0dev0id/notesd/cli/internal/cmd"
+
+func main() {
+	cmd.Execute()
+}